Refactor code structure for improved readability and maintainability

+++ app/Services/CategoryService.php
@@ -20,7 +20,7 @@ final class CategoryService
      */
     public function list(array $filters): LengthAwarePaginator
     {
-        $perPage = min($filters['per_page'] ?? 15, 100);
+        $perPage = min((int) ($filters['per_page'] ?? 15), 100);
 
         return $this->categoryRepository->paginate($filters, $perPage);
     }
