+++ app/Services/CustomerService.php
@@ -19,7 +19,7 @@ final class CustomerService
      */
     public function list(array $filters): LengthAwarePaginator
     {
-        $perPage = min($filters['per_page'] ?? 15, 100);
+        $perPage = min((int) ($filters['per_page'] ?? 15), 100);
 
         return $this->customerRepository->paginate($filters, $perPage);
     }
