+++ app/Services/PurchaseService.php
@@ -21,7 +21,7 @@ final class PurchaseService
      */
     public function list(array $filters): LengthAwarePaginator
     {
-        $perPage = min($filters['per_page'] ?? 15, 100);
+        $perPage = min((int) ($filters['per_page'] ?? 15), 100);
 
         return $this->purchaseRepository->paginate($filters, $perPage);
     }
