+++ app/Services/StockMovementService.php
@@ -21,7 +21,7 @@ final class StockMovementService
      */
     public function list(array $filters): LengthAwarePaginator
     {
-        $perPage = min($filters['per_page'] ?? 15, 100);
+        $perPage = min((int) ($filters['per_page'] ?? 15), 100);
 
         return $this->stockMovementRepository->paginate($filters, $perPage);
     }
